ARM: Determine base and size of malloc space from SDRAM
Now that we know how where to find memory we no longer depend on hardcoded
values. If no malloc size has been specified in Kconfig set up the malloc
space based on the available SDRAM. We use the SDRAM base up to TEXT_BASE
for malloc and unused memory where we put the kernel later.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>