ARM at91sam926x: provide reset function for boards with lowlevel init
Some Atmel boards use a reset() function provided by the architecture.
Add a reset function for the board instead and call the architecture
specific init function from there. While at it, test if we are running
from SDRAM and skip the init function in this case. This makes the configs
safe for both first and second stage.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>