ARM: disable HAVE_CONFIGURABLE_MEMORY_LAYOUT
HAVE_CONFIGURABLE_MEMORY_LAYOUT was first meant as a feature, now it's a
feature to remove it. barebox on ARM now completely uses the memory passed
in from the lowlevel code and configures the malloc area and stack space
during runtime making it obsolete to hardcode these values. With the
removal of HAVE_CONFIGURABLE_MEMORY_LAYOUT barebox no longer requests the
SDRAM region for the stack automatically, so add code to request the region
manually.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>