state: add framework for persistent state handling
Boards often have the need to store variables in persistent memory.
The constraints are often different from what the regular environment
can do:
- compact binary format to make it suitable for small EEPROMs/MRAMs
- atomic save/restore of the whole variable set
- redundancy

This adds a framework to describe, access, store and restore a set of
variables. A state variable set can be fully described in a devicetree
node. This node could be part of the regular devicetree blob or it could
be an extra devicetree solely for the state. The state variable set
contains variables of different types and a place to store the variable
set.

== Backends ==

Currently two backends exist. The raw backend is a very compact format
consisting of a magic value for identification, the raw values and a
CRC.  Two copies are maintained for making sure that during update the
storage device still contains a valid state. The dtb backend stores the
state as a devicetree binary blob. This is exactly the original
devicetree description of the state itself, but additionally contains
the actual values of the variables. Unlike the raw state backend the dtb
state backend can describe itself.

== Frontend ==

As frontend a state instance is a regular barebox device which has
device parameters for the state variables. With this the variables
can be accessed like normal shell variables. The 'state' command is
used to save/restore a state to the backend device.

== Variables ==

State variables have a type. Currenty supported types are: enum32,
uint32 and MAC Address. Fixed length strings are planned but not
implemented. Variable length strings are not planned.

== Example ==

This is a devicetree blob describing a state with a single variable.
The state is stored in eeprom using the raw backlend:

	state: state@0 {
		magic = <0x27031977>;
		compatible = "barebox,state";
		backend-type = "raw";
		backend = &eeprom, "partname:state";

		myval@0 {
			type = "u32";
			default = <0x1>;
		};
	};

After initializing the variable can be accessed with $state.myval.
state -s stores the state to eeprom.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>