Drivers/Net/FEC: be able to delay the phy probing
Phy probing must be delayed at the Arkona Link4 platform, because the phy
is part of the FPGA firmware and only present when the FPGA's firmware is
loaded after a cold start.

This change supports probing the FEC on demand via the command 'detect' and
thus can be used in Barebox's shell scripts.

Signed-off-by: Juergen Beisert <jbe@pengutronix.de>