ARM: am335x bone: Fix Booting from eMMC
Booting a beaglebone black from eMMC is broken since:

commit 0d6392de4ad824a6553c0e3e3e18edef689a7c85
Author: Sascha Hauer <s.hauer@pengutronix.de>
Date:   Thu Jun 5 12:09:07 2014 +0200

    ARM: AM335x: Beaglebone: Use stripped down devicetree for MLO

The stripped down device tree does not have the eMMC device node which
the beaglebone black needs for booting. Add this node to the common dts
file, but keep it disabled. It gets enabled later with a call to
am33xx_of_register_bootdevice() when the system is booted from eMMC.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>