usb: gadget: fastboot: detect device if not present
When a device file is not present when the fastboot gadget is created
then try to detect it before we fail. It may be that we want to write
to a device that we haven't used before.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Applied-Upstream: v2018.11.0, commit:37bc0b2daf75a11fe7ae94c574be6ca2d053126d