Makefile: add buildsystem version variable
Introduce a new variable to store a buildsystem version information.
The exact information that is stored here (git commit, git tag,...) is
defined by the buildsystem.
It is intended to have the possibility to get information about the
exact barebox binary, environment and configuration.

The variable is utilized the same as KERNELVERSION for linux:

   make BUILDSYSTEM_VERSION=$COMMITISH

Via scripts/mkcompile_h this information is injected into the barebox
codebase.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>